Refactor workflow step
