+++ flowserv/controller/serial/engine/base.py
@@ -27,7 +27,7 @@ from flowserv.controller.base import WorkflowController
 from flowserv.controller.serial.engine.runner import exec_workflow
 from flowserv.controller.serial.worker.factory import WorkerFactory
 from flowserv.controller.serial.workflow.result import RunResult
-from flowserv.controller.serial.workflow.step import ContainerStep
+from flowserv.model.workflow.step import ContainerStep
 from flowserv.model.base import RunObject
 from flowserv.model.files.factory import FS
 from flowserv.model.template.base import WorkflowTemplate
@@ -278,7 +278,7 @@ def run_workflow(
         Current workflow state (to access the timestamps)
     output_files: list(string)
         Relative path of output files that are generated by the workflow run
-    steps: list of flowserv.controller.serial.workflow.step.WorkflowStep
+    steps: list of flowserv.model.workflow.step.WorkflowStep
         Steps in the serial workflow that are executed in the given context.
     arguments: dict
         Dictionary of argument values for parameters in the template.
