+++ tests/controller/serial/worker/test_docker_worker.py
@@ -15,7 +15,7 @@ import os
 import pytest
 
 from flowserv.controller.serial.worker.docker import DockerWorker
-from flowserv.controller.serial.workflow.step import ContainerStep
+from flowserv.model.workflow.step import ContainerStep
 
 
 # Test files directory
